refactor(options): extract prompt preview generation
